Reorganized files
- added error handling when content size is 0
