+++ daemon/server/include/getImageSize.h
