+++ daemon/server/include/post_manage.h
