+++ daemon/server/include/reqInfo.h
